requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
PyGithub==2.1.1
dnspython==2.4.2
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
asyncio==3.4.3 